# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tagg@d API",
    "version": "1.0.0",
    "description": "API documentation for the Tagg@d backend — a crypto payment platform for Africa."
  },
  "servers": [
    {
      "url": "http://localhost:5002/api/v2",
      "description": "Current version (v2)"
    },
    {
      "url": "http://localhost:5002/api/v1",
      "description": "Deprecated version (v1)"
    },
    {
      "url": "http://localhost:5002",
      "description": "Development Server (unversioned root)"
    },
    {
      "url": "https://taggedpay.xyz/api/v2",
      "description": "Production (v2)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
...
  },
  "paths": {
...
    "/api/api-keys/{keyId}": {
      "get": {
        "summary": "Get a specific API key by ID",
        "tags": [
          "API Keys"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "keyId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "API key ID"
          }
        ],
        "responses": {
          "200": {
            "description": "API key details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "scopes": {
                          "type": "string"
                        },
                        "is_active": {
                          "type": "boolean"
                        },
                        "last_used_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API key not found"
          }
        }
      },
      "patch": {
        "summary": "Update an API key",
        "description": "Update the name, scopes, IP whitelist, or rotation interval of an existing API key.",
        "tags": [
          "API Keys"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "keyId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 100
                  },
                  "scopes": {
                    "type": "string"
                  },
                  "ipWhitelist": {
                    "type": "string"
                  },
                  "rotationIntervalDays": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 365
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "API key not found"
          }
        }
      },
      "delete": {
        "summary": "Revoke an API key",
        "description": "Permanently revoke an API key. All requests using this key will be rejected.",
        "tags": [
          "API Keys"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "keyId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key revoked successfully"
          },
          "404": {
            "description": "API key not found"
          }
        }
      }
    },
...
    "/admin/rate-limits/settings": {
      "get": {
        "summary": "Get current rate limit settings",
        "description": "Returns the configured rate limit thresholds for each endpoint tier.",
        "tags": [
          "Rate Limit Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Rate limit configuration",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "description": "Rate limit settings per endpoint"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/admin/rate-limits/users/{userId}": {
      "get": {
        "summary": "Get a user's rate limit status",
        "description": "Returns the current rate limit counters and remaining quota for a specific user.",
        "tags": [
          "Rate Limit Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User rate limit status"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/admin/rate-limits/users/{userId}/tier": {
      "put": {
        "summary": "Update a user's rate limit tier",
        "description": "Change a user's rate limit tier (e.g., from free to premium) to adjust their quota.",
        "tags": [
          "Rate Limit Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tier"
                ],
                "properties": {
                  "tier": {
                    "type": "string",
                    "enum": [
                      "free",
                      "basic",
                      "premium",
                      "enterprise"
                    ],
                    "example": "premium"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User tier updated"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/admin/rate-limits/api-keys/{keyId}": {
      "get": {
        "summary": "Get an API key's rate limit status",
        "tags": [
          "Rate Limit Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "keyId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key rate limit status"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/admin/rate-limits/api-keys/{keyId}/rate-limit": {
      "put": {
        "summary": "Update an API key's rate limit configuration",
        "tags": [
          "Rate Limit Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "keyId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "maxRequests": {
                    "type": "integer",
                    "example": 5000
                  },
                  "windowMs": {
                    "type": "integer",
                    "example": 3600000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key rate limit updated"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/admin/rate-limits/violations": {
      "get": {
        "summary": "Get rate limit violations log",
        "description": "Returns a log of recent rate limit violations across all users and API keys.",
        "tags": [
          "Rate Limit Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Violations log"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}